Add support for DidChangeConfiguration
This allows the client to change the configuration on the fly.
As well as replace the need for a config file.

Clients send a `DidChangeConfiguration` notification to update the
configuration. Only supplied values will be overwritten.
Supported values are:
  1. `Languages`
  2. `RootMarkers`
  3. `Commands`
  4. `LogLevel`
  5. `LintDebounce`

In addition, because the capabilities are now dynamic, The client can
specify the server capabilities on initialization with
`InitializeParams`.
Supported values are:
  1. `Completion`
  2. `Hover`
  3. `CodeAction`
  4. `DocumentSymbol`
  5. `DocumentFormatting`

LSP spec:
https://microsoft.github.io/language-server-protocol/specification.html#workspace_didChangeConfiguration
https://microsoft.github.io/language-server-protocol/specification.html#initialize