Prevent multiple linters running for the same URI
For long running linters, there was a build up LSP requests that did not
clear until all linters had finished. Now the linters are running in a
separate goroutine with a context that is canceled whenever a new
request for the same URI comes in. This means in effect that the latest
linter instance is the one returning a result.