Fix accidental forkbomb.
exec.CommandContext() creates a Cmd that will kill its process when its
associated Context is cancelled. Unfortunately, that kill signal
is only sent to the parent process of that Cmd, which in this case
was `sh`. Child processes can continue to run and use CPU and memory.

If your configured linter takes a while to run and uses a substantial
amount of memory (ex: eslint w/ TypeScript), and you have the default
lint configuration of linting on textDocument/didChange, then
it's quite easy to type quickly enough to forkbomb your laptop and
cause it ti run out of its many gigabytes of RAM.

This fixes that by making sure to kill the spawned lint process and
all of its children.

Note: This doesn't *only* fix the case where we run a command
within the `sh` hard-coded in the source, like this:

```
sh
+-- my_lint_command
```

But also fixes the case where the lint comand you call delegates to
other commands like this:

```
sh
+-- make
  +-- pnpm
    +-- eslint
```